requirement.txt
streamlit
transformers
sentence-transformers
faiss-cpu
langchain-text-splitters
numpy
torch